+++ source/utils/types.js
@@ -1,0 +1,6 @@
+// @flow
+import type { ElementRef } from 'react';
+
+export type ReactElementRef<ElementType = HTMLElement> = {
+  current: null | ElementRef<ElementType>
+}