+++ stories/index.js
@@ -5,6 +5,7 @@ import './Bubble.stories';
 import './Button.stories';
 import './Checkbox.stories';
 import './Header.stories';
+import './InfiniteScroll.stories';
 import './Input.stories';
 import './Layout.stories';
 import './LoadingSpinner.stories';